+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to Audit.NET and its extensions will be documented in this f
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [12.3.0] - 2018-08-20
+### Modified
+- Audit.Core: Changing the exception serialization to include the stacktrace and inner exceptions.
+
 ## [12.2.2] - 2018-08-15
 ### Modified
 - Audit.WebApi, Audit.WebApi.Core: Adding AuditIgnoreAttribute for controller, actions and arguments.